+++ CHANGELOG.md
@@ -16,6 +16,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com),
 and this project adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased]
+### Added
+- most task properties can now also be defined as CLI arguments (#23)
+
 ### Changed
 - Removed dependency on java-plugin (#19)